Fix: Resolve implicit-widening-of-multiplication-result in DescriptorAllocation
- m_descriptorSize * offsetInBlock was computed in UINT32, so any
  overflow would truncate before the addition to m_descriptor.ptr
- Cast the first operand to SIZE_T so the multiplication itself
  happens in 64 bits; no realistic overflow in current usage, but
  removes the hazard for free

+++ D3D12Renderer/DescriptorAllocation.cpp
@@ -91,7 +91,7 @@ void DescriptorAllocation::Free()
 D3D12_CPU_DESCRIPTOR_HANDLE DescriptorAllocation::GetDescriptorHandle(UINT32 offsetInBlock) const
 {
     assert(offsetInBlock < m_numHandles);
-    return {m_descriptor.ptr + (m_descriptorSize * offsetInBlock)};
+    return {m_descriptor.ptr + static_cast<SIZE_T>(m_descriptorSize) * offsetInBlock};
 }
 
 UINT32 DescriptorAllocation::GetOffset() const
